refactor(native): folderize major native engine source domains
