+++ src/native/engine/ai/controller/controller_system_lifecycle.c
