+++ src/native/engine/ai/navigation/navigation.c
