+++ src/native/engine/ai/npc/npc_merchant.c
