+++ src/native/engine/ai/wildlife/wildlife_controller.c
