+++ src/native/engine/render/registry/actor_mesh_registry.c
