+++ src/native/engine/runtime/camera/follow/camera_follow_policy.c
