+++ src/native/engine/runtime/parallel/parallel_contract.c
@@ -1,18 +1,18 @@
 #include "parallel_contract.h"
 
 static const RuntimeParallelBudget kRuntimeParallelBudgets[] = {
-    {"ai.dispatch.update", "src/native/engine/ai/controller_system_dispatch.c", 16, 0.60},
-    {"ai.dispatch.signal", "src/native/engine/ai/controller_system_dispatch.c", 28, 0.58},
-    {"ai.dispatch.detach", "src/native/engine/ai/controller_system_dispatch.c", 41, 0.58},
+    {"ai.dispatch.update", "src/native/engine/ai/controller/controller_system_dispatch.c", 16, 0.60},
+    {"ai.dispatch.signal", "src/native/engine/ai/controller/controller_system_dispatch.c", 28, 0.58},
+    {"ai.dispatch.detach", "src/native/engine/ai/controller/controller_system_dispatch.c", 41, 0.58},
     {"tools.asset.generate.layers", "src/native/engine/tools/application/banana_asset_generation_service.c", 58, 0.62},
-    {"physics.dynamics.integrate", "src/native/engine/physics/dynamics.c", 54, 0.45},
-    {"physics.world_collision.sweep", "src/native/engine/physics/world_collision.c", 15, 0.40},
-    {"world.query.nearby", "src/native/engine/world/world_query.c", 26, 0.52},
+    {"physics.dynamics.integrate", "src/native/engine/physics/dynamics/dynamics.c", 54, 0.45},
+    {"physics.world_collision.sweep", "src/native/engine/physics/world/world_collision.c", 15, 0.40},
+    {"world.query.nearby", "src/native/engine/world/query/world_query.c", 26, 0.52},
     {"runtime.controller_sync.update", "src/native/engine/runtime/controller/controller_sync.c", 55, 0.50},
     {"runtime.controller_sync.entity_sync", "src/native/engine/runtime/controller/controller_sync.c", 75, 0.45},
-    {"render.mesh_terrain.vertices", "src/native/engine/render/mesh_terrain.c", 30, 0.55},
-    {"render.mesh_terrain.normals", "src/native/engine/render/mesh_terrain.c", 74, 0.55},
-    {"render.renderer.instance_upload", "src/native/engine/render/renderer.c", 259, 0.64},
+    {"render.mesh_terrain.vertices", "src/native/engine/render/mesh/mesh_terrain.c", 30, 0.55},
+    {"render.mesh_terrain.normals", "src/native/engine/render/mesh/mesh_terrain.c", 74, 0.55},
+    {"render.renderer.instance_upload", "src/native/engine/render/renderer/renderer.c", 259, 0.64},
     {"runtime.physics_abi.sync", "src/native/engine/runtime/abi/physics/physics_abi.c", 39, 0.48},
     {"tools.asset.cellular.step", "src/native/engine/tools/domain/banana_asset_cellular.c", 41, 0.60},
     {"tools.asset.wfc.entropy", "src/native/engine/tools/domain/banana_asset_wfc.c", 107, 0.63},
