+++ src/native/engine/ui/inventory/ui_inventory.c
