+++ src/native/engine/win32_dx12_poc/app/main.c
@@ -1,11 +1,11 @@
 #if defined(_WIN32)
 
-#include "../engine.h"
-#include "scene_overlay.h"
-#include "scene_flow.h"
-#include "objective_policy.h"
-#include "../render/backend.h"
-#include "../render/backend_dx12.h"
+#include "../../engine.h"
+#include "../scene_overlay.h"
+#include "../scene_flow.h"
+#include "../objective_policy.h"
+#include "../../render/backend.h"
+#include "../../render/backend_dx12.h"
 
 #include <math.h>
 #include <stdlib.h>
