+++ src/native/engine/win32_dx12_poc/overlay/scene_overlay.c
@@ -1,6 +1,6 @@
-#include "scene_overlay.h"
+#include "../scene_overlay.h"
 
-#include "../engine.h"
+#include "../../engine.h"
 
 #include <stdio.h>
 #include <string.h>
