+++ src/native/engine/win32_dx12_poc/scene/scene_flow.c
@@ -1,4 +1,4 @@
-#include "scene_flow.h"
+#include "../scene_flow.h"
 
 static void clear_result(BananaPocSceneFlowResult *result)
 {
