+++ src/native/engine/world/entity/entity.c
